Determine RAM size and fixed data locations at runtime
Hardcoded addresses are finally no longer used! Unused parts of the
interrupt vector table are used to store data and pointers to data
used at runtime. The RAM size is calculated on boot and passed to the
booted operating system. data.asm contains the locations and sizes of
any writable data required. Bump version to 0.10.0.